ZEN-4408 Move common files to genflow-common
New package crg.common.codegen will house common classes and files; SCG
and OAG will continue to have their own projects utilizing the common
core.
