+++ modules/gentemplates/swagger-codegen/pom.xml
@@ -34,8 +34,8 @@
             <artifactId>commons-csv</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.openapitools</groupId>
-            <artifactId>openapi-generator</artifactId>
+            <groupId>com.reprezen.genflow</groupId>
+            <artifactId>genflow-common</artifactId>
         </dependency>
     </dependencies>
     <build>
